made comment changes; changed file names and class names; structure changes; packaged files
