+++ .idea/compiler.xml
@@ -12,7 +12,7 @@
   </component>
   <component name="JavacSettings">
     <option name="ADDITIONAL_OPTIONS_OVERRIDE">
-      <module name="pickitup" options="-parameters" />
+      <module name="pickitup" options="" />
     </option>
   </component>
 </project>
