--- src/main/java/com/example/pickitup/service/BusinessService.java
@@ -1,11 +1,0 @@
-package com.example.pickitup.service;
-
-import org.springframework.stereotype.Service;
-
-@Service //Marks this class as a service component for business logic
-public class BusinessService {
-    //Returns a welcome message that can be used in the UI
-    public String getWelcomeMessage() {
-        return "Welcome to Pick-It-Up!";
-    }
-}
